🔥 폴더 재구성
각자 이름의 폴더 생성
